back/requirements.txt
fastapi==0.126.0
uvicorn[standard]==0.34.0
sqlalchemy==2.0.46
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
python-multipart==0.0.20
pandas==2.3.3
numpy==2.0.2
scikit-learn==1.6.1
openai==2.14.0
httpx==0.28.1
email-validator==2.3.0